Move helpers, remove offset from post json
